requirements.txt
pillow
numpy
requests
spotipy
RGBMatrixEmulator
python-dateutil
keyboard